pyk/*, test/*: make new smodule proof for new proof files
