+++ src/pyk/kcfg/__init__.py
@@ -1,5 +1,4 @@
 from .explore import KCFGExplore
 from .kcfg import KCFG
-from .proof import AllPathReachabilityProof
 from .show import KCFGShow
 from .tui import KCFGViewer
