+++ src/pyk/proof/reachability.py
@@ -1,14 +1,12 @@
 from __future__ import annotations
 
 import logging
-from abc import abstractmethod
-from enum import Enum
 from typing import TYPE_CHECKING, Any, Callable, Dict, Final, Iterable, Optional, Type, TypeVar
 
+from ..kcfg import KCFG, KCFGExplore
 from ..prelude.ml import mlAnd
 from ..utils import shorten_hashes
-from .explore import KCFGExplore
-from .kcfg import KCFG
+from .proof import Proof, ProofStatus
 
 if TYPE_CHECKING:
     from pathlib import Path
@@ -19,40 +17,6 @@ if TYPE_CHECKING:
 T = TypeVar('T', bound='Proof')
 
 _LOGGER: Final = logging.getLogger(__name__)
-
-
-class ProofStatus(Enum):
-    PASSED = 'passed'
-    FAILED = 'failed'
-    PENDING = 'pending'
-
-
-class Proof:
-    _PROOF_TYPES: Final = {'AllPathReachabilityProof'}
-
-    @classmethod
-    def _check_proof_type(cls: Type[T], dct: Dict[str, Any], expected: Optional[str] = None) -> None:
-        expected = expected if expected is not None else cls.__name__
-        actual = dct['type']
-        if actual != expected:
-            raise ValueError(f'Expected "type" value: {expected}, got: {actual}')
-
-    @classmethod
-    @abstractmethod
-    def from_dict(cls: Type[Proof], dct: Dict[str, Any]) -> Proof:
-        proof_type = dct['type']
-        if proof_type in Proof._PROOF_TYPES:
-            return globals()[proof_type].from_dict(dct)
-        raise ValueError(f'Expected "type" value in: {Proof._PROOF_TYPES}, got {proof_type}')
-
-    @property
-    @abstractmethod
-    def status(self) -> ProofStatus:
-        ...
-
-    @abstractmethod
-    def to_dict(self) -> Dict[str, Any]:
-        ...
 
 
 class AllPathReachabilityProof(Proof):
