+++ src/tests/integration/kcfg/test_imp.py
@@ -8,9 +8,10 @@ import pytest
 from pyk.cterm import CSubst, CTerm
 from pyk.kast.inner import KApply, KSequence, KSort, KToken, KVariable, Subst
 from pyk.kast.manip import get_cell
-from pyk.kcfg import KCFG, AllPathReachabilityProof
+from pyk.kcfg import KCFG
 from pyk.prelude.kint import intToken
 from pyk.prelude.ml import mlAnd, mlBottom, mlEqualsFalse, mlEqualsTrue
+from pyk.proof import AllPathReachabilityProof
 
 from ..utils import KCFGExploreTest
